MicroML

Documentation for MicroML.

MicroML.AbstractML

MicroML.AbstractMLASTNode

MicroML.AbstractMLType

MicroML.AbstractMLVal

MicroML.MLApp

MicroML.MLBool

MicroML.MLId

MicroML.MLIf

MicroML.MLInt

MicroML.MLLambda

MicroML.MLOp

MicroML.MLToken

MicroML.MLTypeVar

MicroML.TypeEquation

MicroML.Compiler

MicroML.Lexer

MicroML.Parser

MicroML.apply_unifier

MicroML.assign_typenames

MicroML.eval

MicroML.generate_equations

MicroML.get_expression_type

MicroML.occurs_check

MicroML.show_type_assignment

MicroML.unify

MicroML.unify_equations

MicroML.unify_variable

MicroML.visit_children

MicroML.AbstractML — Type

AbstractML

Top type of MicroML

MicroML.AbstractMLASTNode — Type

AbstractMLASTNode <: AbstractML

attribute

type :: String Used by the type inference algorithm.

children :: Vector Used by passes that traverse the AST. Each concrete node class lists the sub-nodes it has as children.

method

visit_children(n::AbstractMLASTNode, f::Function)

show(io::IOBuffer, n::AbstractMLASTNode)

MicroML.AbstractMLType — Type

AbstractMLType <: AbstractML

attribute

name :: String

method

string(t::AbstractMLType)

isequal

MicroML.AbstractMLVal — Type

AbstractMLVal <: AbstractMLASTNode

attribute

type :: String

children :: Vector

value :: String

method

string(n::AbstractMLVal)

MicroML.MLApp — Type

Application of a function to a sequence of arguments. func is a node, args is a sequence of nodes.

MicroML.MLBool — Type

Bool Constant

MicroML.MLId — Type

Identifier

MicroML.MLIf — Type

if ... then ... else ... expression.

MicroML.MLInt — Type

Int Constant

MicroML.MLLambda — Type

lambda [args] -> expr

MicroML.MLOp — Type

Binary operation between expressions.

MicroML.MLToken — Type

A simple Token structure. Contains the token type, value and position.

MicroML.MLTypeVar — Type

A type variable.

MicroML.TypeEquation — Type

TypeEquation <: AbstractML

A type equation between two types: left and right. original is the original AST node from which this equation was derived, for debugging.

MicroML.Compiler — Method

Compiler()

A Compiler() object.

Property

attribute

p :: Parser

equations

symtab

code

method

compile(source::String)

interpret()

MicroML.Lexer — Function

Lexer(rules=RULES; skip_whitespace=rrue)

A simple regex-based lexer/tokenizer.

Property

attribute

buf :: String

pos :: Int

method

start(buf::String)

token()

tokens()

peek()

MicroML.Parser — Method

Parser()

Parser for micro-ML.

Usage:

p = Parser()
decl = p.parse(<some micro-ML code>)
# decl is now an ast.Decl node

Property

attribute

lexer

token :: MLToken

method

parse(source::String, should_terminate::Bool=True)

MicroML.apply_unifier — Method

apply_unifier(typ::AbstractMLType, subst::Subst) :: Subst

Applies the unifier subst to typ. Returns a type where all occurrences of variables bound in subst were replaced (recursively); on failure returns None.

MicroML.assign_typenames — Function

assign_typenames(node, symtab::SymTab=SymTab())

Assign typenames to the given AST subtree and all its children.

Symtab is the initial symbol table we can query for identifiers found throughout the subtree. All identifiers in the subtree must be bound either in symtab or in lambdas contained in the subtree. This function updates the type property on the AST nodes it visits.

MicroML.eval — Function

Eval ML code.

MicroML.generate_equations — Function

generate_equations(node, type_equations::Vector=[])

Generate type equations from node and place them in typeequations. Prior to calling this functions, node and its children already have to be annotated with _type, by a prior call to assigntypenames.

MicroML.get_expression_type — Function

get_expression_type(
    expr, 
    subst::Subst,
    rename_types::Bool=true
) :: AbstractMLType

Finds the type of the expression given a substitution.

If rename_types is True, renames all the type vars to be sequential characters starting from 'a', so that 't5 -> t3' will be renamed to 'a -> b'. These names are less cluttery and also facilitate testing.

Note: expr should already be annotated with assign_typenames.

MicroML.occurs_check — Method

occurs_check(v::AbstractMLType, typ::AbstractMLType, subst::Subst) :: Bool

Does the variable v occur anywhere inside typ? Variables in typ are looked up in subst and the check is applied recursively.

MicroML.show_type_assignment — Method

show_type_assignment(node)

Show a type assignment for the given subtree, as a table. Returns a string that shows the assigmnent.

MicroML.unify — Method

unify(typ_x, typ_y, subst)

Unify two types typx and typy, with initial subst. Returns a subst (map of name->Type) that unifies typx and typy, or None if they can't be unified. Pass subst={} if no subst are initially known. Note that {} means valid (but empty) subst.

MicroML.unify_equations — Method

unify_equations(eqs::Vector) :: Subst

Unifies all type equations in the sequence eqs. Returns a substitution (most general unifier).

MicroML.unify_variable — Method

unify_variable(v::AbstractMLType, typ::AbstractMLType, subst::Subst) :: Subst

Unifies variable v with type typ, using subst. Returns updated subst or None on failure.

MicroML.visit_children — Method

Visit all children with a function that takes a child node.